Only set cflags/libs when download is successful

+++ get-libgit2-linux.sh
@@ -5,8 +5,8 @@ else
 OLDOPENSSL=$("${R_HOME}/bin/Rscript" -e 'cat(openssl::openssl_config()$version)' | grep "OpenSSL 1.0")
 if [ $? -eq 0 ] && [ "$OLDOPENSSL" ]; then
 URL="https://r-lib.github.io/gert/libgit2-1.1.0.x86_64_legacy-linux.tar.gz"
-PKG_CFLAGS="-DSTATIC_LIBGIT2 -I${PWD}/libgit2/include"
-PKG_LIBS="-L${PWD}/libgit2/lib -lgit2 -lrt -lpthread -lssh2 -lssl -lcrypto -ldl -lpcre -lz"
+TMP_CFLAGS="-DSTATIC_LIBGIT2 -I${PWD}/libgit2/include"
+TMP_LIBS="-L${PWD}/libgit2/lib -lgit2 -lrt -lpthread -lssh2 -lssl -lcrypto -ldl -lpcre -lz"
 else
 OPENSSL3=$("${R_HOME}/bin/Rscript" -e 'cat(openssl::openssl_config()$version)' | grep "OpenSSL 3")
 if [ $? -eq 0 ] && [ "$OPENSSL3" ]; then
@@ -15,10 +15,13 @@ echo "Found OpenSSL3"
 else
 URL="https://autobrew.github.io/archive/x86_64_linux/libgit2-1.4.2-x86_64_linux.tar.gz"
 fi
-PKG_CFLAGS="-DSTATIC_LIBGIT2 -I${PWD}/libgit2-1.4.2-x86_64_linux/include"
-PKG_LIBS="-L${PWD}/libgit2-1.4.2-x86_64_linux/lib -lgit2 -lrt -lpthread -lssh2 -lssl -lcrypto -ldl"
+TMP_CFLAGS="-DSTATIC_LIBGIT2 -I${PWD}/libgit2-1.4.2-x86_64_linux/include"
+TMP_LIBS="-L${PWD}/libgit2-1.4.2-x86_64_linux/lib -lgit2 -lrt -lpthread -lssh2 -lssl -lcrypto -ldl"
 fi
-"${R_HOME}/bin/R" -s -e "curl::curl_download('$URL','bundle.tar.gz')"
-tar xzf bundle.tar.gz && rm -f bundle.tar.gz
-HAVE_STATIC_LIBGIT2=TRUE
+if "${R_HOME}/bin/R" -s -e "curl::curl_download('$URL','bundle.tar.gz')"; then
+  tar xzf bundle.tar.gz && rm -f bundle.tar.gz
+  HAVE_STATIC_LIBGIT2=TRUE
+  PKG_CFLAGS="$TMP_CFLAGS"
+  PKG_LIBS="$TMP_LIBS"
+fi
 fi
